Move core/types.h to model to prevent a core -> local -> core dependency cycle
